# openapi.yaml
...
openapi: 3.0.3
info:
  title: CFactory API
  version: 0.1.0
  summary: Agentic control-tower cockpit over the PARR pipeline.
  description: >-
    Read-first, advise-and-confirm cockpit over the three Factory services
    (PFactory · AIFactory · TFactory). Ingests normalized completion events,
    hydrates a correlation-keyed WorkItem store, exposes board/rollup/anomaly
    reads, an LLM copilot, and a propose → confirm → execute action flow with a
    tamper-evident audit trail.
  license:
    name: MIT OR GPL-3.0
servers:
  - url: http://localhost:3111
    description: Local backend (CFACTORY_BACKEND_PORT, default 3111)
  - url: https://cfactory.freundcloud.com
    description: Hosted cockpit
...
paths:
...
  /api/cards:
    get:
      tags: [Cards]
      summary: List planning cards (the backlog), highest priority first
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/CardStatus" } }
        - { name: milestone, in: query, schema: { type: string } }
        - { name: assignee, in: query, schema: { type: string } }
        - { name: tier, in: query, schema: { $ref: "#/components/schemas/CardTier" } }
      responses:
        "200":
          description: This tenant's cards.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  cards:
                    type: array
                    items: { $ref: "#/components/schemas/Card" }
    post:
      tags: [Cards]
      summary: Create a planning card (requires the `write` scope)
      description: >-
        Omit `card_key` to have the next `FCT-<n>` assigned for the tenant.
        Audited.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Card" }
      responses:
        "201":
          description: The created card.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "409": { description: The tenant already holds that card_key. }
  /api/cards/import:
    post:
      tags: [Cards]
      summary: Import the repository's EXISTING issues as cards (requires `write`)
      description: >-
        RFC-0020 §3.6. Lists the configured project's issues through the shared
        git-provider protocol (so GitHub, GitLab and Azure DevOps alike) and
        upserts each as a card. Imported cards land in `backlog` — a closed issue
        in `done` — and **never** in `ready`: `ready` + a tier is the dispatch
        trigger, so importing a repo full of `factory:low` issues must not fire a
        build per issue. Idempotent (an upsert against a unique
        `(tenant_id, issue_ref)` index), pull requests are never imported, and it
        is **poll-based, not live** — there is no webhook receiver, so an issue
        filed since the last run appears on the next one. Audited.
      parameters:
        - { name: project, in: query, schema: { type: string },
            description: "Project path to import from; defaults to the tenant's configured intake project." }
        - { name: full, in: query, schema: { type: boolean, default: false },
            description: "Ignore the last-synced watermark and re-read every issue." }
      responses:
        "200":
          description: >-
            The import tally. `ok` is false (still 200) when the provider was
            unreachable or unconfigured.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  project: { type: string }
                  imported: { type: integer }
                  updated: { type: integer }
                  skipped:
                    type: integer
                    description: Issues whose card was soft-deleted locally, or which failed to map.
                  seen: { type: integer }
                  truncated:
                    type: boolean
                    description: CFACTORY_IMPORT_MAX capped the run — the board is NOT the whole backlog.
                  import_max: { type: integer }
                  incremental: { type: boolean }
                  live:
                    type: boolean
                    description: Always false. Import is a poll; webhooks remain deferred.
                  last_synced_at:
                    type: string
                    format: date-time
                    nullable: true
                    description: >-
                      The incremental CURSOR (newest issue updated_at seen, minus a
                      clock-skew overlap) — NOT when the read happened. It does not
                      move on a repository nobody has touched; use `polled_at`.
                  polled_at:
                    type: string
                    format: date-time
                    nullable: true
                    description: >-
                      When this pass read the provider successfully — what "last
                      synced" means to a human. Null on a failed pass.
                  reason: { type: string }
...
  /api/cards/{card_key}:
    parameters:
      - { name: card_key, in: path, required: true, schema: { type: string } }
    get:
      tags: [Cards]
      summary: Get one planning card
      responses:
        "200":
          description: The card.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "404": { description: No card for the key (in this tenant). }
    patch:
      tags: [Cards]
      summary: Update a card — also how it MOVES (status) and is REPRIORITISED (priority)
      description: >-
        Partial update: only fields present in the body are applied. Requires
        the `write` scope. Audited.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Card" }
      responses:
        "200":
          description: The updated card.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "404": { description: No card for the key (in this tenant). }
    delete:
      tags: [Cards]
      summary: Delete a planning card (requires the `write` scope). Audited.
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  card_key: { type: string }
                  deleted: { type: boolean }
        "404": { description: No card for the key (in this tenant). }
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <key>`. Scoped keys configured via
        CFACTORY_API_KEYS (`<key>:read,write;...`). Open when no keys are set.
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    CardStatus:
      type: string
      enum: [backlog, ready, in_progress, blocked, done]
    CardTier:
      type: string
      description: RFC-0011 difficulty tier.
      enum: [low, medium, hard]
...
    Card:
      type: object
      description: >-
        RFC-0019 §3.1 planning card. Lives in its own `cards` table, NOT on
        work_items (whose rows the reconcile/prune machinery rewrites and
        deletes from upstream polling). `correlation_key` is null while the
        card is only planned and joins work_items once it enters the factory.
      properties:
        card_key: { type: string, description: "Stable human id, e.g. FCT-42. Unique per tenant." }
        tenant_id: { type: string }
        title: { type: string }
        description:
          type: string
          nullable: true
          description: >-
            Free-form markdown body (RFC-0020 §3.6). Where an imported issue's
            body lands. MIRRORED — the host owns it, like the title — and
            deliberately NOT parsed into `acceptance_criteria`, which would
            fabricate the statements the factory verifies against.
        acceptance_criteria: { type: array, items: { type: string } }
        status: { $ref: "#/components/schemas/CardStatus" }
        priority: { type: integer, description: "Lower = higher priority." }
        tier: { $ref: "#/components/schemas/CardTier" }
        assignee:
          type: string
          nullable: true
          description: A human handle or a factory runtime (e.g. "aifactory").
        milestone: { type: string, nullable: true }
        correlation_key: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...